Improve the implementation of CHECK() macro.
CHECK_EQ() currently does not work with smart pointers and nullptr.
This is an attempt to improve the situation by improving the
EagerEvaluator template. As a side benefit, the new implementation
no longer uses macros to generate template specializations.

The primary motivation of this change is to improve compatibility
with logging macros from google3 (Google's internal monorepo) and
make it easier to export code that was developed there to Android.

Bug: 151119760
Test: atest --host libbase_test

Change-Id: I9c957d5ddf2afb87a88304b1e4d7ceebea0fa2f8
Signed-off-by: smokey18 <kmboplays@gmail.com>